make parse_file always use len arg as length
Since the rest of the function always assumes that the value of the
len argument contains the length of lines, there is no reason to have
a backup case, if the rest of the function would break anyway.
Furthermore, len should ALWAYS have a valid value anyway, since it is
not optional.