add clear_location function and handle more fatals
Since some fatal errors are strictly internal we should not print
any of the current locatoin info. Therefore a clear_location function
was added to make this possibly, even from the user code.

Furthermore some of the internal errors that can occur in the error
module, will now report themselves with the report_fatal function,
instead of just crashing.

Lastly the report_fatal function will now always report the error
message even if no location is present. Again this is because the
location info might not be relevant in all cases.