rework compilation function
The array returned from read_file now also contains entries for
non-code lines. This is because we need those to display the correct
line number if an error occurs during parse_file. This also means that
parse_file has to recount the amount of lines, since some of them
might not be actual code lines.

Lastly the responsibility of adding the file header and end code
has been moved to build_file, since it fits better there.