print warning when using pin i/o w/o power config
If the CHECK or SET command is used without having specified a vin
and ground we issue a warning. This is because many chips will not
function as expected if they do have any power connected.

Since this warning will also clutter the output of a lot of test
cases we also add the -w flag to all the program calls to avoid
unnecessary warnings in the output.